move utils to common folder
