+++ components/google_drive/actions/list-files/list-files.mjs
@@ -1,5 +1,5 @@
 import googleDrive from "../../google_drive.app.mjs";
-import { getListFilesOpts } from "../../utils.mjs";
+import { getListFilesOpts } from "../../common/utils.mjs";
 
 export default {
   key: "google_drive-list-files",
